+++ client/Packages/com.beamable/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Helper menu option for updating Realm configuration for GPGS integration
 
+### Changed
+- `AccountManagementConfiguration` no longer overiddes `BEAMABLE_GPGS` define symbol. 
+
 ### Fixed
 - CID/PID Mismatch error message was too big for Unity popup. Now it uses a Beamable Custom Editor Window. [3933](https://github.com/beamable/BeamableProduct/issues/3933)